Resurrect old `Signature` object (with internal names),
It seems that immutable identifiers for functions and their arguments are still
very useful in a few places. So the goal is to make internal names be very
limited throughout the system to make it easier to debug most objects passed
around, and only use them when necessary.

Also, some simple tools for renaming